Fix to avoid non-termination due to changes in plug-in interface.
This avoids non-termination in some common cases.  For details,
see the comment just above `isVarEq`.

In short: if a plug-in generates something which is equivalent
(but not syntactically the same) to something already in the inert set,
and GHC notices that fact, we go into a loop, because the plug-in just
keeps generating the new fact, and GHC keeps ignoring it.

The current fix is a work around, but this should really be fixed in
TcInteract in GHC---it can let us know if it disregarded a fact, and
nothing else changed, the way it used to in an earlier version.