Add more examples, and a tricky bit that does not work.
We end up in this situation:

(2 <= a) => (a ~ ?n + 1)

This is interesting because while we know that ?n will always exist,
we don't know *what* it will be.  If we used subtraction, we could
define `?n ~ (a - 1)`, which in this case is safe because we know that
a is at least two.

We can start doing that, once the story with partial functions, such
as subtraction, clears up a bit more.